requirements.txt
flask
pandas
requests
asyncio
python-nmap
smtplib